Create EX
I have a problem when I try to do cases 2 and 3. The variable "activo" works like a switch when that struct is "created" (I know I am using static memory, but the  assignment makes me do it this way). The problem is that the variable "componente.activo = 0" dosent save that variable in "componente[i].activo" and since i am calling the "void" function on the main with a "for" and inside this function there is  the "componente[i]" element on it, wouldn't this save the position where the "i"  variable is with the value that the "loop" gives to it?